+++ VueJs.js
@@ -4,7 +4,6 @@
  */
 const merge = require("webpack-merge");
 const VueLoaderPlugin = require("vue-loader/lib/plugin");
-const MiniCssExtractPlugin = require("mini-css-extract-plugin");
 
 /**
  * Prepares the webpack config for a vue js environment
@@ -56,59 +55,5 @@ module.exports = class VueJs {
 
 		// Disalbe component loader
 		context.currentAppConfig.componentLoader = false;
-
-		// Add style loaders
-		this._applyStyleLoaders(context);
-	}
-
-	isComponentEnabled(state, component) {
-		// Disalbe default style loaders
-		if (component === "SassLoader" || component === "LessLoader") return false;
-		return state;
-	}
-
-	/**
-	 * Applies additional style loader required for vue to work propperly with sass and less
-	 * @param {module.ConfigBuilderContext} context
-	 * @private
-	 */
-	_applyStyleLoaders(context) {
-		context.webpackConfig.module.rules.push({
-			test: /\.scss$/,
-			use: [
-				context.isProd ? MiniCssExtractPlugin.loader : "vue-style-loader",
-				"css-loader",
-				"sass-loader"
-			]
-		});
-		context.webpackConfig.module.rules.push({
-			test: /\.sass$/,
-			use: [
-				context.isProd ? MiniCssExtractPlugin.loader : "vue-style-loader",
-				"css-loader",
-				{
-					loader: "sass-loader",
-					options: {
-						indentedSyntax: true
-					}
-				}
-			],
-
-		});
-		context.webpackConfig.module.rules.push({
-			test: /\.css$/,
-			use: [
-				context.isProd ? MiniCssExtractPlugin.loader : "vue-style-loader",
-				"css-loader"
-			]
-		});
-		context.webpackConfig.module.rules.push({
-			test: /\.less$/,
-			use: [
-				context.isProd ? MiniCssExtractPlugin.loader : "vue-style-loader",
-				"css-loader",
-				"less-loader"
-			]
-		});
 	}
 };