+++ plugins/VueJsEnvPlugin.js
@@ -19,4 +19,42 @@ module.exports = class VueJsEnvPlugin {
 		if(context.environment !== "vuejs" || type !== "esLint") return pattern;
 		return basePattern;
 	}
+
+	filterLoaderConfig(config, type, context){
+		// Only inject the vue loader in a development context
+		if(context.isProd) return config;
+
+		// Rewrite sass loader
+		if(type === "sassLoader"){
+			config.use.forEach((v, k) => {
+				if(typeof v === "string") v = {loader: v};
+				if(typeof v.loader === "undefined") return;
+
+				// Inject vue style loader
+				if(v.loader.match(/mini-css-extract-plugin/))
+					config.use[k] = "vue-style-loader";
+			});
+		}
+
+		// Rewrite less loader
+		if(type === "lessLoader"){
+			config.use.forEach((v, k) => {
+				if(typeof v === "string") v = {loader: v};
+				if(typeof v.loader === "undefined") return;
+
+				// Inject vue style loader
+				if(v.loader.match(/mini-css-extract-plugin/))
+					config.use[k] = "vue-style-loader";
+			});
+		}
+	}
+
+	customSassLoaderFileExtensionFallback(ext, stylesheetPath, resourceQuery){
+		// Help out with single file components
+		if(ext !== "vue") return ext;
+		if(resourceQuery.match(/&lang=sass/)) return "sass";
+		if(resourceQuery.match(/&lang=scss/)) return "scss";
+		if(resourceQuery.match(/&type=style/) && resourceQuery.match(/&type=.*?/) === null) return "css";
+		return ext;
+	}
 };